requirements.txt
langchain==0.0.228
gpt4all==1.0.3
chromadb==0.3.26
llama-cpp-python==0.1.68
urllib3==2.0.3
PyMuPDF==1.22.5
python-dotenv==1.0.0
unstructured==0.8.0
extract-msg==0.41.5
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.65.0
sentence_transformers
#chromadb-client
termcolor
chromaviz # or pip install git+https://github.com/mtybadger/chromaviz/